'styles' directory is moved under the 'features' directory
