+++ src/features/styles/libsbmlnetwork_styles.h
