+++ src/libsbmlnetwork_sbmldocument_render.cpp
@@ -5,7 +5,7 @@
 #include "libsbmlnetwork_render.h"
 #include "libsbmlnetwork_render_helpers.h"
 #include "libsbmlnetwork_layout_helpers.h"
-#include "styles/libsbmlnetwork_styles.h"
+#include "features/styles/libsbmlnetwork_styles.h"
 #include "features/error_log/libsbmlnetwork_error_log.h"
 #include "features/defaults/libsbmlnetwork_defaults_render.h"
